refactor: refactor code according to SOLID
