+++ src/pages/PromptDetail/Messages.jsx
@@ -2,15 +2,10 @@ import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
 import { Box } from '@mui/material';
 import IconButton from '@mui/material/IconButton';
 import List from '@mui/material/List';
-import ListItem from '@mui/material/ListItem';
-import MenuItem from '@mui/material/MenuItem';
-import Select from '@mui/material/Select';
 import Snackbar from '@mui/material/Snackbar';
-import TextField from '@mui/material/TextField';
-
 import { styled } from '@mui/material/styles';
 import { Fragment, useCallback, useEffect, useState } from 'react';
-import { DragDropContext, Draggable, Droppable } from "react-beautiful-dnd";
+import { DragDropContext, Droppable } from "react-beautiful-dnd";
 import { useDispatch, useSelector } from 'react-redux';
 
 import {
@@ -19,15 +14,12 @@ import {
   StyledAccordionSummary,
   StyledTypography
 } from '@/components/BasicAccordion';
-import ArrowDownIcon from '@/components/Icons/ArrowDownIcon';
-import CopyIcon from '@/components/Icons/CopyIcon';
-import DeleteIcon from '@/components/Icons/DeleteIcon';
-import MoveIcon from '@/components/Icons/MoveIcon';
 import PlusIcon from '@/components/Icons/PlusIcon';
 
 import Alert from '@/components/Alert';
 import { PROMPT_PAYLOAD_KEY, ROLES } from '@/pages/PromptDetail/constants.js';
 import { actions } from '@/reducers/prompts';
+import MessageInput from './MessageInput';
 
 const AddButton = styled(IconButton)(() => (`
   width: 1 rem;
@@ -41,74 +33,6 @@ const AddButton = styled(IconButton)(() => (`
   background: rgba(255, 255, 255, 0.10); 
   color: white;
 `));
-
-const MessageContainer = styled(ListItem)(() => `
-  display: flex;
-  height: 9.9125rem;
-  padding: 0.625rem;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 0.325rem;
-  align-self: stretch;
-
-  border-radius: 0.5rem;
-  border: 1px solid #3B3E46;
-
-  :not(:last-child) {
-    margin-bottom: 1rem;
-  }
-`);
-
-const MessageToolbar = styled(Box)(() => `
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  align-self: stretch;
-`);
-
-const StyledSelect = styled(Select)(() => `
-  display: flex;
-  height: 1.88rem;
-  padding: 0.25rem 0rem;
-  align-items: center;
-  gap: 0.625rem;
-  & .MuiOutlinedInput-notchedOutline {
-    border-width: 0px;
-  }
-  & .Mui-focused .MuiOutlinedInput-notchedOutline {
-    border: 0px solid white
-  }
-  & .MuiOutlinedInput-input {
-    padding: 0.25rem 0 0.5rem
-  }
-  & .MuiSelect-icon {
-    right: 0px;
-  }
-  fieldset{
-    border: none !important;
-    outline: none !important;
-  };
-`)
-
-const StyledArrowDownIcon = styled(ArrowDownIcon)(() => `
-  fill: white;
-`)
-
-const ButtonsContainer = styled(Box)(() => `
-display: flex;
-align-items: flex-start;
-gap: 0.5rem;
-`);
-
-const StyledTextField = styled(TextField)(() => `
-  flex: 1 0 0;
-  color: #FFF;
-
-  font-size: 0.875rem;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 1.375rem; /* 157.143% */
-`);
 
 const MessageList = styled(List)(() => `
   width: 100%;
@@ -138,74 +62,6 @@ const StrictModeDroppable = ({ children, ...props }) => {
   return <Droppable direction="vertical" {...props}>{children}</Droppable>;
 };
 
-const Message = ({ index, id, role, content, onChangeRole, onDelete, onChangeContent, onCopy }) => {
-  const onSelectRole = useCallback((event) => {
-    onChangeRole(event.target.value);
-  }, [onChangeRole]);
-
-  const onClickDelete = useCallback(
-    () => {
-      onDelete();
-    },
-    [onDelete],
-  );
-
-  const onChangeInput = useCallback(
-    (event) => {
-      onChangeContent(event.target.value)
-    },
-    [onChangeContent],
-  );
-
-  return (
-    <Draggable key={id} draggableId={id} index={index}>
-      {(provided) => (
-        <MessageContainer
-          ref={provided.innerRef}
-          {...provided.draggableProps}
-
-        >
-          <MessageToolbar>
-            <StyledSelect
-              value={role}
-              onChange={onSelectRole}
-              displayEmpty
-              IconComponent={StyledArrowDownIcon}
-            >
-              <MenuItem value={ROLES.Assistant}>Assistant</MenuItem>
-              <MenuItem value={ROLES.System}>System</MenuItem>
-              <MenuItem value={ROLES.User}>User</MenuItem>
-            </StyledSelect>
-            <ButtonsContainer>
-              <IconButton onClick={onClickDelete}>
-                <DeleteIcon sx={{ fontSize: '1.13rem' }} />
-              </IconButton>
-              <IconButton disabled={!content} onClick={onCopy}>
-                <CopyIcon sx={{ fontSize: '1.13rem' }} />
-              </IconButton>
-              <IconButton {...provided.dragHandleProps} >
-                <MoveIcon sx={{ fontSize: '1.13rem' }} />
-              </IconButton>
-            </ButtonsContainer>
-          </MessageToolbar>
-          <StyledTextField
-            autoFocus
-            fullWidth
-            id="standard-multiline-static"
-            label=""
-            value={content}
-            multiline
-            rows={4}
-            onChange={onChangeInput}
-            variant="standard"
-            placeholder="Input message here"
-            InputProps={{ disableUnderline: true }}
-          />
-        </MessageContainer>
-      )}
-    </Draggable>
-  )
-}
 
 const Messages = () => {
   const dispatch = useDispatch();
@@ -336,7 +192,7 @@ const Messages = () => {
                       {
                         messages.map((message, index) => {
                           return (
-                            <Message
+                            <MessageInput
                               key={message.id}
                               id={message.id}
                               index={index}
